Add new quote and update README.md
Prior to this commit, the README did not reflect the intended usage of
this project as a learning tool. This commit updates the README to
describe how students should use this repo to practice creating pull
requests on Github and includes a sample quote by Teddy Roosevelt in
the quotes/ directory.